requirements.txt
python-docx-1.1.2 
whoosh-2.7.4
fhirpy
pandas